bumped cypress version and made changes to support the upgrade
